+++ quiz-game/src/assets/index.js
@@ -1,0 +1,9 @@
+import darkMode from './dark-mode.svg';
+import lightMode from './light-mode.svg';
+import questions from './questions.svg';
+
+export default {
+    darkMode,
+    lightMode,
+    questions
+}
