+++ quiz-game/src/components/navigation/Route.jsx
@@ -1,0 +1,3 @@
+export function Route({children}) {
+    return children
+}
