+++ quiz-game/src/components/navigation/index.js
@@ -1,0 +1,4 @@
+export { Router } from "./Router";
+export { Route } from "./Route";
+export { Routes } from "./Routes";
+export { Link } from "./Link";
