+++ quiz-game/src/context/RouterContext.jsx
@@ -1,0 +1,8 @@
+import {createContext} from "react";
+
+export const RouterContext = createContext(
+    {
+        location: "",
+        push: () => {}
+    }
+)
