+++ quiz-game/src/hooks/useRouterContext.js
@@ -1,0 +1,6 @@
+import {useContext} from "react";
+import {RouterContext} from "@context/RouterContext.jsx";
+
+export function useRouterContext() {
+    return useContext(RouterContext)
+}
