+++ quiz-game/src/pages/MainPage.jsx
@@ -1,0 +1,12 @@
+import styles from './pages.module.css';
+import Game from "@components/Game/index.js";
+
+function MainPage() {
+    return (
+        <>
+            <Game />
+        </>
+    )
+}
+
+export default MainPage;
