+++ quiz-game/src/pages/NotFound/NotFoundPage.jsx
@@ -1,0 +1,9 @@
+function NotFoundPage() {
+    return (
+        <>
+
+        </>
+    )
+}
+
+export default NotFoundPage;
