+++ quiz-game/src/pages/QuizPage.jsx
@@ -1,0 +1,11 @@
+import styles from './pages.module.css';
+
+function QuizPage() {
+    return (
+        <>
+
+        </>
+    )
+}
+
+export default QuizPage;
