+++ quiz-game/src/pages/ScorePage.jsx
@@ -1,0 +1,11 @@
+import styles from './pages.module.css';
+
+function ScorePage() {
+    return (
+        <>
+
+        </>
+    )
+}
+
+export default ScorePage;
